+++ azure-functions-maven-plugin/CHANGELOG.md
@@ -16,7 +16,7 @@ All notable changes to the "Maven Plugin for Azure Function" will be documented 
 
 ## 1.6.0
 - Support Java 11 Azure Functions (Preview). [#1052](https://github.com/microsoft/azure-maven-plugins/issues/1052)
-- Support specify Azure environment in configuration. [PR#1095](https://github.com/microsoft/azure-maven-plugins/pull/1095)
+- Support specify Azure environment for auth method 'azure_auth_maven_plugin'. [PR#1095](https://github.com/microsoft/azure-maven-plugins/pull/1095)
 
 ## 1.5.0
 - Support creating application insights while creating or updating function apps. [PR#1074](https://github.com/microsoft/azure-maven-plugins/pull/1074)